Moved most of the code from main into the game structure. Continued fixing all the resulting errors. Changed the type from f32 to f64 with all the physics components

+++ src/physics.rs
@@ -4,13 +4,13 @@ use specs::{Component, DenseVecStorage, ParJoin, ReadStorage, System, VecStorage
 
 #[derive(Debug, Component)]
 #[storage(VecStorage)]
-pub struct Position(pub Point2<f32>);
+pub struct Position(pub Point2<f64>);
 
 #[derive(Debug, Component)]
-pub struct Velocity(pub Vector2<f32>);
+pub struct Velocity(pub Vector2<f64>);
 
 #[derive(Debug, Component)]
-pub struct Acceleration(pub Vector2<f32>);
+pub struct Acceleration(pub Vector2<f64>);
 
 pub struct PhysicsSystem;
